Refactor into three projects: binlog (library), binlog.mcp (MCP server), binlog.cli (CLI)
- Create binlog/ class library with all analysis tools and infrastructure
- Slim binlog.mcp/ down to just Program.cs and *Extensions.cs files
- Create binlog.cli/ with CLI commands (without MCP server command)
- Remove args from server.json transport (MCP server now launched directly)
- Make BinlogJsonContext and BinlogJsonOptions public for cross-project access
- Update solution file to include all three projects
