+++ binlog.cli/Program.cs
@@ -1,0 +1,6 @@
+using Binlog.MCP.Cli;
+using System.CommandLine;
+
+CliRunner.RegisterCallbacks();
+var rootCommand = CliCommands.BuildRootCommand();
+return await rootCommand.InvokeAsync(args);
