+++ binlog.cli/binlog.cli.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net10.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <PublishSingleFile>true</PublishSingleFile>
+    <PublishTrimmed>true</PublishTrimmed>
+    <PackAsTool>true</PackAsTool>
+    <PackageId>baronfel.binlog.cli</PackageId>
+    <Description>A CLI tool for analyzing and investigating MSBuild binlogs.</Description>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <Authors>Chet Husk</Authors>
+    <PackageTags>msbuild;binlog;cli</PackageTags>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="System.CommandLine" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="../binlog/binlog.csproj" />
+  </ItemGroup>
+</Project>
