+++ binlog.mcp/.mcp/server.json
@@ -9,8 +9,7 @@
       "identifier": "baronfel.binlog.mcp",
       "version": "0.0.13",
       "transport": {
-        "type": "stdio",
-        "args": ["mcp"]
+        "type": "stdio"
       }
     }
   ],
