+++ binlog/Features/AnalyzerAnalysis/AnalyzerParser.cs
