+++ binlog/Features/AnalyzerAnalysis/Models.cs
