+++ binlog/Features/BuildAnalysis/BinlogPrompts.cs
