+++ binlog/Features/DiagnosticAnalysis/GetDiagnosticsTool.cs
