+++ binlog/Features/DiagnosticAnalysis/Models.cs
