+++ binlog/Features/EvaluationAnalysis/GetEvaluationPropertiesByNameTool.cs
