+++ binlog/Features/EvaluationAnalysis/GetEvaluationPropertiesTool.cs
