+++ binlog/Features/ProjectAnalysis/ExpensiveProjectsTool.cs
