+++ binlog/Features/ProjectAnalysis/GetProjectBuildTimeTool.cs
