+++ binlog/Features/ProjectAnalysis/ListProjectsTool.cs
