+++ binlog/Features/ProjectAnalysis/Models.cs
