+++ binlog/Features/ProjectAnalysis/ProjectBuildTimeCache.cs
