+++ binlog/Features/SearchAnalysis/Models.cs
