+++ binlog/Features/SearchAnalysis/SearchBinlogTool.cs
