+++ binlog/Features/TargetAnalysis/TargetInfoTools.cs
