+++ binlog/Features/TimelineAnalysis/TimelineCache.cs
